Crate snowcap_main

Modules

example_topologies

transient_violation

Structs

CommandLineArguments
This is the binary to use the runtime systen esily. This program will generate the topology and the reconfiguration scenario (based on the options provided), synthesize a reconfiguration order and perform this order on a network simulated inside GNS3 using FRRouting.

Enums

MainCommand

NetworkSelection
This is the binary to use the runtime systen esily. This program will generate the topology and the reconfiguration scenario (based on the options provided), synthesize a reconfiguration order and perform this order on a network simulated inside GNS3 using FRRouting.

Scenario

Functions

check_config

custom_scenario

get_topo

main

topology_zoo_scenario