Enum snowcap_main::MainCommand

pub(crate) enum MainCommand {
    Synthesize {
        use_tree: bool,
        network: NetworkSelection,
    },
    Optimize {
        use_tree: bool,
        network: NetworkSelection,
    },
    Runtime {
        network: NetworkSelection,
        persistent_gns_project: bool,
        random_sequence: bool,
        at_once: bool,
        seed: Option<u64>,
        json_filename: Option<String>,
    },
    Bencher {
        network: NetworkSelection,
        args: BencherArguments,
    },
    TransientViolation {
        gml_file: String,
        seed: u64,
        n_iter: usize,
        n_seeds: usize,
        reverse: bool,
        num_threads: Option<usize>,
    },
    CustomOperation {
        n_iter: usize,
        variant: usize,
    },
}

Variants

Synthesize

Perform the migration synthesis

Fields of Synthesize

use_tree: bool

Use the tree strategy instead of the more complex one

network: NetworkSelection

Type of measurement to perform

Optimize

Perform the migration synthesis using soft policies

Fields of Optimize

use_tree: bool

Use the tree strategy instead of the more complex one

network: NetworkSelection

Type of measurement to perform

Runtime

Perform the migration synthesis and the Runtime simulation

Fields of Runtime

network: NetworkSelection

Type of measurement to perform

persistent_gns_project: bool

Leave the gns3 project open after quitting this project

random_sequence: bool

Use a random sequence for the reconfiguration

at_once: bool

Apply all modifiers at once, without monitoring the network

seed: Option<u64>

Seed for the random sequence (if used)

json_filename: Option<String>

Store the result summary in a json file

Bencher

Run the Bencher

Fields of Bencher

network: NetworkSelection

Type of measurement to perform

args: BencherArguments

Bencher Arguments

TransientViolation

Verify transient condition and violations

Fields of TransientViolation

gml_file: String

GML file to use for topology zoo

seed: u64

Random seed, to get reproducable networks

n_iter: usize

Number of iterations for approximating probability of transient violation

n_seeds: usize

Number of different seeds to evaluate, repeating the same experiment multiple times

reverse: bool

perform the modification in reverse (remove the eBGP peer)

num_threads: Option<usize>

Number of parallel execution units

CustomOperation

Custom method

Fields of CustomOperation

n_iter: usize

Number of iterations for computing

variant: usize

Variant of the scenario to compute the violation